GitHub Edit [https://github.com/Door43/ulb-en/commit/90765e1a0ece27cb11fefb221c90e84ae28b10e3]: Judges 2:2 change wording
Judges 2:2 change wording from You must make no treaty with those who live... to what you see because the same word (berith) is used. There is a special significance to berith or covenant with religious connotations. That's why there is the reference in the next sentence to altars. [Hanker10: Hanker10]