GitHub Edit [https://github.com/Door43/ulb-en/commit/b4a27187530fe1b74e18fda4940314a56315827b]: Ezk 3:20
No version I consulted, nor NICOT, makes I set a stumbling bloc the
result of the preceding clause. All make it part of the protasis, and
only he will die as the apodosis. I have followed the placement of
the zaqeps and 'athnach for the logic of the second sentence. [hmw3: hmw3]